refactor: pairRoomController 테스트를 PairRoomAcceptance 테스트로 변경
